Split the semantic issue into a base class and the actual class
