+++ src/Semantic/Issue.php
@@ -9,85 +9,8 @@ use SPFLib\Record;
 /**
  * Class that represent a semantic issue reported by SemanticVamidator.
  */
-class Issue
+class Issue extends AbstractIssue
 {
-    /**
-     * Semantic warning code: too many terms that involve DNS lookups.
-     *
-     * @var int
-     */
-    public const CODE_TOO_MANY_DNS_LOOKUPS = 1;
-
-    /**
-     * Semantic warning code: 'all' should be the last mechanism.
-     *
-     * @var int
-     */
-    public const CODE_ALL_NOT_LAST_MECHANISM = 2;
-
-    /**
-     * Semantic warning code: both the 'all' mechanism and the 'redirect' modifier are present ('redirect' will be ignored).
-     *
-     * @var int
-     */
-    public const CODE_ALL_AND_REDIRECT = 3;
-
-    /**
-     * Semantic warning code: the 'ptr' mechanism should be avoided (it's slow, expensive, unreliable).
-     *
-     * @var int
-     */
-    public const CODE_SHOULD_AVOID_PTR = 4;
-
-    /**
-     * Semantic warning code: the 'redirect' and 'exp' modifiers should appear after mechanisms.
-     *
-     * @var int
-     */
-    public const CODE_MODIFIER_NOT_AFTER_MECHANISMS = 5;
-
-    /**
-     * Semantic warning code: the 'redirect' and 'exp' modifiers can't appear more that 1 time.
-     *
-     * @var int
-     */
-    public const CODE_DUPLICATED_MODIFIER = 6;
-
-    /**
-     * Semantic warning code: unknown modifier specified.
-     *
-     * @var int
-     */
-    public const CODE_UNKNOWN_MODIFIER = 7;
-
-    /**
-     * Semantic warning code: a macro-string contains the "p" (validated domain) macro-letter, which should be avoided (it's slow, expensive, unreliable).
-     *
-     * @var int
-     */
-    public const CODE_SHOULD_AVOID_VALIDATED_DOMAIN_MACRO = 8;
-
-    /**
-     * Issue level: notice (can be ignored).
-     *
-     * @var int
-     */
-    public const LEVEL_NOTICE = 1;
-
-    /**
-     * Issue level: warning (should be fixed).
-     *
-     * @var int
-     */
-    public const LEVEL_WARNING = 2;
-
-    /**
-     * Issue level: fatal (must be fixed).
-     *
-     * @var int
-     */
-    public const LEVEL_FATAL = 3;
-
     /**
      * The affected record.
      *
@@ -96,39 +19,19 @@ class Issue
     private $record;
 
     /**
-     * The code of the issue (the value of one of the Issue::CODE_... constants).
-     *
-     * @var int
-     */
-    private $code;
-
-    /**
-     * The issue description.
-     *
-     * @var string
-     */
-    private $description;
-
-    /**
-     * The issue level (the value of one of the Issue::LEVEL_... constants).
-     *
-     * @var int
-     */
-    private $level;
-
-    /**
      * Initialize the instance.
-     *
-     * @param string $term the term that wasn't recognized
      */
     public function __construct(Record $record, int $code, string $description, int $level)
     {
+        parent::__construct($code, $description, $level);
         $this->record = $record;
-        $this->code = $code;
-        $this->description = $description;
-        $this->level = $level;
     }
 
+    /**
+     * {@inheritdoc}
+     *
+     * @see \SPFLib\Semantic\AbstractIssue::__toString()
+     */
     public function __toString(): string
     {
         $level = $this->getLevelDescription();
@@ -142,41 +45,5 @@ class Issue
     public function getRecord(): Record
     {
         return $this->record;
-    }
-
-    /**
-     * Get the code of the issue (the value of one of the Issue::CODE_... constants).
-     */
-    public function getCode(): int
-    {
-        return $this->code;
-    }
-
-    /**
-     * Get the issue description.
-     */
-    public function getDescription(): string
-    {
-        return $this->description;
-    }
-
-    /**
-     * Get the issue level (the value of one of the Issue::LEVEL_... constants).
-     */
-    public function getLevel(): int
-    {
-        return $this->level;
-    }
-
-    protected function getLevelDescription(): string
-    {
-        switch ($this->getLevel()) {
-            case static::LEVEL_NOTICE:
-                return 'notice';
-            case static::LEVEL_WARNING:
-                return 'warning';
-            case static::LEVEL_FATAL:
-                return 'fatal';
-        }
     }
 }
